sdk/src/cameras, sdk/include/aditof: Update Ini Params in DepthCompute
added method to update m_depthINIDataMap when parameters are changed
on ADSD3500
when a parameter value is changed, the data structure to be passed to
DepthCompute is updated with the changed value. The new structure is then
used to reinitialise DepthCompute
added init_dc boolean variable to adsd3500SetXX methods. When init_dc is
false, the reinitialisation of DepthCompute is skipped.

Signed-off-by: Cristina Suteu <cristina.suteu@analog.com>